refactor: Улучшение файловой структуры проекта и минорные улучшения docker-compose файлов
