+++ docker-compose.prod.yml
@@ -1,11 +1,9 @@
 
 services:
   api:
-    command: uvicorn src.app.main:app --host 0.0.0.0 --port 8000
+    command: uvicorn app.main:app --host 0.0.0.0 --port 8000
     ports:
       - "8000:8000"
 
   worker:
-    command: celery -A src.app.worker.celery_app worker --loglevel=info
-
-
+    command: celery -A app.worker.celery_app worker --loglevel=info
