+++ docker-compose.yml
@@ -1,6 +1,9 @@
 
 x-base-app: &base-app
   build: .
+  working_dir: /app/src
+  environment:
+    - PYTHONPATH=/app
   env_file:
     - .env
   restart: unless-stopped
