+++ src/app/main.py
@@ -1,8 +1,7 @@
 from fastapi import FastAPI
 
-from src.app.api.v1.endpoints.jobs import router as jobs_router
-from src.app.api.v1.endpoints.process import router as process_router
-from src.app.db.session import Base, engine  
+from src.app.api.main import main_router
+from src.db.session import Base, engine  
 
 
 Base.metadata.create_all(bind=engine)
@@ -10,8 +9,7 @@ Base.metadata.create_all(bind=engine)
 
 def create_app():
     app = FastAPI()
-    app.include_router(jobs_router, prefix="/v1")
-    app.include_router(process_router, prefix="/v1")
+    app.include_router(main_router)
     return app
